refactor code to better support multi operating system functionality in the future
