+++ osq-exts/tables/crowdstrikefalconagent/README.md
@@ -1,0 +1,5 @@
+This table reads CrowdStrike Falcon Agent Information into an OSQuery Table.
+
+Currently, macOS is supported with the CrowdStrike Agent 6.18+
+
+Eventual Goal is to have this a multi operating system extension, with support for Windows and Linux.
